+++ scripts/main.js
@@ -1,8 +1,9 @@
 import { registerHumanCultures } from "./content/human-cultures.js";
 import { registerAncestryNamesI } from "./content/ancestry-names-i.js";
+import { registerAncestryNamesII } from "./content/ancestry-names-ii.js";
 
 export const MODULE_ID = "pf2e-npc-forge-names-inner-sea";
-export const VERSION = "0.2.0";
+export const VERSION = "0.3.0";
 
 let registered = false;
 
@@ -18,6 +19,7 @@ export function registerContent(api) {
   }
   registerHumanCultures(api);
   registerAncestryNamesI(api);
+  registerAncestryNamesII(api);
   registered = true;
   console.log(`${MODULE_ID} | Registered Names of the Inner Sea ${VERSION}`);
   return true;